Added ability to set default value for envs (#46)
Now it's possible to set default value for environment variables.
This value is used if the variable isn't set in the shell.

Example:
```yaml
  elements:
    - images:
      - name: genesis-base
        format: qcow2
        envs:
          - A
          - B=foo
```

The `A` variable has default behavior. If it's not set in the shell an
empty value is set to it. But for the `B` variable the string "foo" is
set in this case.

Signed-off-by: Anton Kremenetsky <anton.kremenetsky@gmail.com>